requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2024.2.2
channels==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
Django==4.2.6
django-cors-headers==4.3.1
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
mysqlclient==2.2.3
openai==1.12.0
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
